pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "livekit-flows"
version = "0.1.0"
description = "A library for building declarative agent flows with LiveKit."
readme = "README.md"
license = "MIT"
authors = [
    {name = "Mateusz Kulpa", email = "mateusz.kulpa@outlook.com"},
]
maintainers = [
    {name = "Mateusz Kulpa", email = "mateusz.kulpa@outlook.com"},
]
keywords = [
    "livekit",
    "ai",
    "conversational-ai",
    "voice",
    "agents",
    "flow",
    "workflow",
    "declarative",
    "real-time",
    "webrtc"
]
requires-python = ">=3.12"
dependencies = [
    "livekit-agents>=1.3.5",
    "pydantic>=2.12.5",
    "aiohttp>=3.13.2",
    "jinja2>=3.1.6",
    "pyyaml>=6.0.3",
    "jsonschema>=4.25.1",
]

[project.urls]
Homepage = "https://github.com/mateuszkulpa/livekit-flows"
Repository = "https://github.com/mateuszkulpa/livekit-flows"

[tool.hatch.build.targets.wheel]
packages = ["livekit_flows"]

[tool.uv]
default-groups = ["dev", "examples"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-aiohttp>=1.1.0",
    "ruff>=0.14.7",
    "ty>=0.0.1a29",
    "pytest-repeat>=0.9.4",
]
examples = [
    "livekit-agents>=1.3.5",
    "livekit-plugins-openai>=1.3.5",
    "livekit-plugins-cartesia>=1.3.5",
    "livekit-plugins-deepgram>=1.3.5",
    "livekit-plugins-silero>=1.3.5",
    "python-dotenv>=1.2.1",
]
